Move JWT and auth to common package
